moved get_token_model() to models, updated the docs
