+++ knox/models.py
@@ -1,4 +1,6 @@
+from django.apps import apps
 from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
 from django.db import models
 from django.utils import timezone
 
@@ -45,3 +47,21 @@ class AbstractAuthToken(models.Model):
 class AuthToken(AbstractAuthToken):
     class Meta:
         swappable = 'KNOX_TOKEN_MODEL'
+
+
+def get_token_model():
+    """
+    Return the AuthToken model that is active in this project.
+    """
+
+    try:
+        return apps.get_model(knox_settings.TOKEN_MODEL, require_ready=False)
+    except ValueError:
+        raise ImproperlyConfigured(
+            "TOKEN_MODEL must be of the form 'app_label.model_name'"
+        )
+    except LookupError:
+        raise ImproperlyConfigured(
+            "TOKEN_MODEL refers to model '%s' that has not been installed"
+            % knox_settings.TOKEN_MODEL
+        )
